Updated the header navigation bar css
