--- client/modules/products.js
@@ -1,15 +1,0 @@
-import {getCategoryList, getProducts, productsLoad, showCart } from './products/products.js';
-
-window.onload = (event) => {
-    productsLoad();
-    getCategoryList();
-    getProducts();
-    let cart = document.querySelector('#cartAmount .counter');
-    cart.textContent = JSON.parse(localStorage.getItem("cart")).length || 0;
-}
-
-let cartButton = document.getElementsByClassName('cart');
-cartButton[0].addEventListener('click', () => {
-    showCart();
-});
-
