+++ client/products-page.html
@@ -11,34 +11,38 @@
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css">
   <link href="https://fonts.googleapis.com/css2?family=Dosis&display=swap" rel="stylesheet">
-  <script src="./modules/styles/bootstrap.min.js"></script>
+  <script src="./js/styles/bootstrap.min.js"></script>
 </head>
 
 <body>
-  <nav class="navbar navbar-expand-lg navbar-light bg-light navigation-head">
-    <div class="container-fluid container-style">
-      <div class="navbar-brand home-logo">
-        <img class="logo" src="../../../static/images/logo.png" alt="" width="80" height="24">
-      </div>
-      <div class="collapse navbar-collapse" id="navbarSupportedContent">
-        <ul class="navbar-nav home-links me-auto mb-2 mb-lg-0">
-          <li class="nav-item">
-            <a class="nav-link active" aria-current="page" href="./index.html">Home</a>
-          </li>
-          <li class="nav-item">
-            <a id='products' class="nav-link" href="#">Products</a>
-          </li>
-        </ul>
-        <div>
-          <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-            <li class="nav-item">
-              <a class="nav-link active" aria-current="page" href="./signin.html">Signin</a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link" href="./register.html">Register</a>
-            </li>
-          </ul>
-          <div class="cart" data-bs-toggle="modal" data-bs-target="#staticBackdrop">
+  <header>
+    <div class="container-fluid header-container">
+      <div class="row header-row">
+        <div class="col-8">
+          <div class="container-fluid">
+            <div class="row ">
+              <div class="col-6 image-col">
+                <img class="img-fluid" src="../../../static/images/logo.png" alt="" width="100%" height="100%">
+              </div>
+              <div class="col-6 home-tabs">
+                <nav class="custom-navbar" aria-labelledby="mainmenu">
+                  <ul class="nav-links nav-btns">
+                    <li class="list-btns"><a class="link-button" href="./index.html">Home</a></li>
+                    <li><a class="link-button" href="#">Products</a></li>
+                  </ul>
+                </nav>
+              </div>
+            </div>
+          </div>
+        </div>
+        <div class="col-4 cart-colunm">
+          <nav class="custom-navbar" aria-labelledby="singup">
+            <ul class="nav-links sign-register-btn">
+              <li class="list-btns"><a class="link-button" href="./signin.html">Sign In</a></li>
+              <li><a class="link-button" href="./register.html">Register</a></li>
+            </ul>
+          </nav>
+          <div class="cart cart-text" data-bs-toggle="modal" data-bs-target="#staticBackdrop">
             <i class="bi bi-cart-fill">
             </i>
             <div id="cartAmount" class="cartAmount"><span class="counter">0</span> Items</div>
@@ -46,7 +50,7 @@
         </div>
       </div>
     </div>
-  </nav>
+  </header>
 
   <!--cart-->
   <div class="modal fade" id="staticBackdrop" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1"
@@ -60,13 +64,13 @@
         </div>
         <div class="modal-body cart-body">
           <div class="container-fluid cart-items-container">
-          
+
           </div>
         </div>
         <div data-bs-dismiss="modal" class="checkout-div">
           <div class="checkout">
             <div class="checkout-text">Proceed to checkout</div>
-            <div class="checkout-amount">Rs 
+            <div class="checkout-amount">Rs
               <span class="amount"></span>
             </div>
           </div>
@@ -77,12 +81,12 @@
 
   <main id="products-main">
     <div class="row products-row">
-      <div class="col-md-2 categories-col">
+      <div class="col-md-3 categories-col">
         <div class="categories-list">
 
         </div>
       </div>
-      <div class="col-md-10 products-col">
+      <div class="col-md-9 products-col">
         <div class="container-fluid">
           <div class="row products-list">
 
@@ -93,7 +97,7 @@
   </main>
 
 
-  <script type="module" src="./modules/products.js"></script>
+  <script type="module" src="./client_index.js"></script>
 </body>
 
 </html>
